refactor : 공통 검증 로직 분리
