+++ src/main/java/domain/Lotto.java
@@ -13,6 +13,12 @@ public class Lotto {
         this.numbers = numbers;
     }
 
+    public static void validateRange(int number) {
+        if (number < LottoConstants.MIN_NUMBER || number > LottoConstants.MAX_NUMBER) {
+            throw new IllegalArgumentException("숫자는 1~45 사이의 정수로 입력해주세요.");
+        }
+    }
+
     public boolean isSameWith(List<Integer> targetNumbers) {
         if(targetNumbers.size() != LottoConstants.NUMBER_COUNT) {
             return false;
@@ -28,16 +34,13 @@ public class Lotto {
         if (numbers.size() != LottoConstants.NUMBER_COUNT) {
             throw new IllegalArgumentException("로또 한 개는 여섯개의 숫자로 이루어져 있어야 합니다.");
         }
-        if (isNotValidRange(numbers)) {
-            throw new IllegalArgumentException("로또 번호는 1~45 사이의 숫자여야 합니다.");
-        }
+        validateRange(numbers);
         if (numbers.size() != new HashSet<>(numbers).size()) {
             throw new IllegalArgumentException("로또의 모든 번호는 서로 중복되지 않아야 합니다.");
         }
     }
 
-    private boolean isNotValidRange(List<Integer> numbers) {
-        return numbers.stream()
-                .anyMatch(n -> n < LottoConstants.MIN_NUMBER || n > LottoConstants.MAX_NUMBER);
+    private void validateRange(List<Integer> numbers) {
+        numbers.forEach(Lotto::validateRange);
     }
 }
