+++ src/main/java/util/Validator.java
@@ -1,6 +1,6 @@
-package view;
+package util;
 
-public class InputValidator {
+public class Validator {
     public static void validateBlank(String value) {
         if (value.isBlank()) {
             throw new IllegalArgumentException("빈 값이 아닌 값을 입력해주세요.");
