+++ src/main/java/view/InputView.java
@@ -1,27 +1,28 @@
 package view;
 
 import java.util.Scanner;
+import util.Validator;
 
 public class InputView {
     private static Scanner sc = new Scanner(System.in);
 
     public String readPurchaseAmount() {
         String purchaseAmountInput = sc.nextLine();
-        InputValidator.validateBlank(purchaseAmountInput);
-        InputValidator.validateIntegerOverflow(purchaseAmountInput);
+        Validator.validateBlank(purchaseAmountInput);
+        Validator.validateIntegerOverflow(purchaseAmountInput);
         return purchaseAmountInput;
     }
 
     public String readWinningNumbers() {
         String inputWinningNumbers = sc.nextLine();
-        InputValidator.validateBlank(inputWinningNumbers);
+        Validator.validateBlank(inputWinningNumbers);
         return inputWinningNumbers;
     }
 
     public String readBonusNumbers() {
         String bonusNumber = sc.nextLine();
-        InputValidator.validateBlank(bonusNumber);
-        InputValidator.validateIntegerOverflow(bonusNumber);
+        Validator.validateBlank(bonusNumber);
+        Validator.validateIntegerOverflow(bonusNumber);
         return bonusNumber;
     }
 }
